requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
opencv-python==4.8.1.78
Pillow==10.1.0
reportlab==4.0.7
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0
email-validator==2.1.0.post1
WTForms==3.1.1
qrcode==8.0
numpy==1.26.4